refactor: split utils/storage.ts into modular structure
Split the 1095-line storage.ts into utils/storage/ directory:
- dbCore.ts: DB constants and openDB function
- projectData.ts: Session management
- analysisCache.ts: File key generation and cache operations
- rules.ts: Analysis rules CRUD
- examples.ts: Example management with category detection
- sessions.ts: Example session management
- stationHistory.ts: Station history (localStorage)
- history.ts: Analysis history management
- issues.ts: Analysis issues tracking
- exportImport.ts: JSON export/import utilities
- index.ts: Re-exports all modules

Original file reduced to 4 lines (re-export only).
