+++ utils/storage/exportImport.ts
@@ -1,0 +1,39 @@
+// Export/Import utilities
+
+import { PhotoRecord } from "../../types";
+import { AnalysisRule } from "./rules";
+
+export const exportDataToJson = (photos: PhotoRecord[]): string => {
+  const dataToExport = photos.map(p => {
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    const { originalFile, ...rest } = p;
+    return rest;
+  });
+  return JSON.stringify(dataToExport, null, 2);
+};
+
+export const importDataFromJson = (jsonStr: string): PhotoRecord[] => {
+  try {
+    const parsed = JSON.parse(jsonStr);
+    if (!Array.isArray(parsed)) throw new Error("Invalid JSON format");
+    return parsed as PhotoRecord[];
+  } catch (e) {
+    console.error("Import failed", e);
+    throw e;
+  }
+};
+
+export const exportRulesToJson = (rules: AnalysisRule[]): string => {
+  return JSON.stringify(rules, null, 2);
+};
+
+export const importRulesFromJson = (jsonStr: string): AnalysisRule[] => {
+  try {
+    const parsed = JSON.parse(jsonStr);
+    if (!Array.isArray(parsed)) throw new Error("Invalid Rules JSON format");
+    return parsed as AnalysisRule[];
+  } catch (e) {
+    console.error("Import rules failed", e);
+    throw e;
+  }
+};
